+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Return non-zero exit code when no packages are found for `search` and `list` commands.
+- Internal performance improvements.
 
 ## [0.10.0] - 2022-10-01
 ### Changed